Finalize namings
