--- src/Scripts/Network/AbstractNetworkStateWrapper.cpp
@@ -1,1 +1,0 @@
-#include "Network/AbstractNetworkStateWrapper.hpp"
